+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - [[#220]](https://github.com/Roboroads/laravel-tinker/issues/220) Add context menu action in tinker consoles to run the console.
 - Replace JB run icon with a green tinker run icon.
+- [[#223]](https://github.com/Roboroads/laravel-tinker/issues/223) Editor font size is now used as font size for the tinker console.
 
 ## [2.5.0] - 2022-12-13